# Ensure that only 1 or 0 is returned when checking if doubles are `NA`